refactor: replace monitor label resolution with unified display label function
- Updated the delete, get, and update monitor mappers to utilize the new getCloudflareMonitorDisplayLabel function for improved consistency in monitor label resolution.
- Removed redundant label resolution functions to streamline the codebase and enhance maintainability.
