+++ web_src/src/pages/workflowv2/mappers/cloudflare/metadata.ts
@@ -28,3 +28,17 @@ export function getCloudflareMonitorDescription(metadata: unknown): string | und
   const raw = m.monitorDescription ?? m.monitor_description;
   return typeof raw === "string" && raw.trim() ? raw.trim() : undefined;
 }
+
+/** Prefer Cloudflare monitor description when node metadata matches the monitor id (see Setup resolver). */
+export function getCloudflareMonitorDisplayLabel(metadata: unknown, monitorId: string): string {
+  const id = monitorId.trim();
+  if (!id) return "-";
+
+  const resolvedId = getCloudflareMonitorId(metadata);
+  const resolvedDesc = getCloudflareMonitorDescription(metadata);
+  if (resolvedId === id && resolvedDesc) {
+    return resolvedDesc;
+  }
+
+  return id;
+}
